testing progress

+++ contracts/NFTDescriptor.sol
@@ -12,20 +12,23 @@ import "./libraries/HexStrings.sol";
 import "./SVGConstants.sol";
 import "./SVGComponents.sol";
 
-
 contract NFTDescriptor is SVGConstants, SVGComponents {
     using SafeMath for uint256;
     using Strings for uint256;
     using HexStrings for uint256;
     using Strings for uint256;
 
-    function amountFormatter(uint256 fixValue, uint256 decimalAmount, uint256 amount
+    function amountFormatter(
+        uint256 fixValue,
+        uint256 decimalAmount,
+        uint256 amount
     ) public pure returns (string memory) {
-        string memory left = (amount / ( 10 ** decimalAmount)).toString();
-        string memory right = ((amount / (10 ** (decimalAmount - fixValue))) % (10 ** (fixValue - 1))).toString();
+        string memory left = (amount / (10**decimalAmount)).toString();
+        string memory right = ((amount / (10**(decimalAmount - fixValue))) %
+            (10**(fixValue - 1))).toString();
         uint256 length = fixValue - bytes(string(right)).length;
-        if(length == 4) return string(left);
-        for(uint i = 0; i < length; i++){
+        if (length == 4) return string(left);
+        for (uint256 i = 0; i < length; i++) {
             right = string(abi.encodePacked("0", right));
         }
         return string(abi.encodePacked(left, ".", right));
@@ -86,8 +89,12 @@ contract NFTDescriptor is SVGConstants, SVGComponents {
             tokenToColorHex(_ownerAddress, 57),
             tokenToColorHex(_ownerAddress, 122)
         ];
-       
-        string memory supportAmount = amountFormatter(5, ERC20(_tokenAddress).decimals(), _supportAmount);
+
+        string memory supportAmount = amountFormatter(
+            5,
+            ERC20(_tokenAddress).decimals(),
+            _supportAmount
+        );
 
         bytes memory dynamicLayer = abi.encodePacked(
             sideText(_tokenId.toString(), "rotate(90 132.5 142.5)", "start"),
